refactor discovery/v2 to discovery/v1
