+++ staging/src/k8s.io/apiserver/pkg/endpoints/discovery/v1/etag.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package v2
+package v1
 
 import (
 	"crypto/sha512"
