+++ staging/src/k8s.io/apiserver/pkg/endpoints/discovery/v1/fake.go
@@ -1,4 +1,4 @@
-package v2
+package v1
 
 import (
 	"context"
